+++ app/[locale]/login/page.tsx
@@ -1,13 +1,13 @@
-import LoginPage from './login-page';
+import Login from './login';
 
-export default function Login({
+export default function LoginPage({
   params: { locale },
 }: {
   params: { locale: string };
 }) {
   return (
     <section className="flex h-screen items-center justify-center">
-      <LoginPage locale={locale} />
+      <Login locale={locale} />
     </section>
   );
 }
